refactor: Moved skill-ability to new directory
