+++ presentation/sheet/item/skill-ability/skill-ability-list-item-viewmodel.mjs
@@ -13,7 +13,7 @@ import InputImageViewModel from "../../../component/input-image/input-image-view
 import InputNumberSpinnerViewModel from "../../../component/input-number-spinner/input-number-spinner-viewmodel.mjs";
 import InputRichTextViewModel from "../../../component/input-rich-text/input-rich-text-viewmodel.mjs";
 import InputTextareaViewModel from "../../../component/input-textarea/input-textarea-viewmodel.mjs";
-import SkillAbility from "../../../../business/ruleset/skill/skill-ability.mjs";
+import SkillAbility from "../../../../business/document/item/skill/skill-ability.mjs";
 import InputTextFieldViewModel from "../../../component/input-textfield/input-textfield-viewmodel.mjs";
 import ButtonRollViewModel from "../../../component/button-roll/button-roll-viewmodel.mjs";
 import ButtonContextMenuViewModel from "../../../component/button-context-menu/button-context-menu-viewmodel.mjs";
